Port and refactor ProjectSettingsProviderFile
Addressed the "share init of properties with base class implementation"
todo by writing a function that sets the parameters that go to the base
constructor before they are passed fully. Also removed the protected
constructor on the parent class.

Addressed the "add Keen Server URL as one of the lines, optionally" by
changing the way the project settings file is read so that it may be
converted to a list which exposes an indexer that is silent on out of
bounds exceptions.

Added the option for using a file that is formatted in JSON and its
corresponding documentation.