Fix another EventCachePortable threading issue.
If an event were dequeued, then another thread came along and added a new event, that new event could take the same name as the dequeued event and overwrite it's backing file, or have it's backing file deleted depending on who gets where first.

This fixes that by not dequeueing the event until the file has been read and deleted.