Enable code coverage reporting for the netstandard tests, maybe.
This requires a couple of tricks to get working. OpenCover doesn't support the new portable PDB format used by default by .NET Core, so DebugType=Full is specified to generate the Windows-style PDB files that OpenCover supports.

There's also an issue with the NUnit3TestAdapter package, where it doesn't include the Mono.Cecil package when DebugType=Full, so that reference is explicitly added as a workaround.